(Closes #11) Auto-indenting in <p> tag causes bug
Unable to understand why text within a <p> tag is autoindented one
layer left. Disabling autoindent and smartindent didn't fix, either.
Instead, changed code to more cleverly find the starting index of
the opening '<', negating any errors.

If the situation is
```
<p>
    This is a <b|>
</p>
```
Pressing <p> still autoindents to this
```
<p>
This is a <b>|</b>
</p>
```
But no longer returns an error when attempting to write the closing
HTML tag.